refactor: 拆分 tools/todo mod 为职责子模块
